requirements-dev.txt
# =============================================================================
# DeltaCrown Tournament Engine - Development Dependencies
# =============================================================================

# Testing Framework
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-xdist==3.5.0
pytest-mock==3.12.0
factory-boy==3.3.0
faker==20.1.0

# Code Quality - Linting
flake8==6.1.0
pylint==3.0.3
pylint-django==2.5.5

# Code Quality - Formatting
black==23.12.1
isort==5.13.2

# Code Quality - Type Checking
mypy==1.7.1
django-stubs==4.2.7
djangorestframework-stubs==3.14.5

# Security
bandit==1.7.6
safety==3.2.0
pip-audit==2.6.1

# Pre-commit Hooks
pre-commit==3.6.0

# Debugging
ipdb==0.13.13
django-debug-toolbar==4.2.0

# Performance Profiling
django-silk==5.0.4

# Load Testing
locust==2.20.0

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# Coverage Reporting
coverage==7.3.3
